pyproject.toml
[project]
name = "handm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "kaggle>=1.7.4.5",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "polars>=1.30.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]

[tool.ruff]
line-length = 120
